+++ finance/estimate_data_utils.py
@@ -6,8 +6,10 @@ from django.utils import timezone
 def calculateGigCost(estimate):
     ret = {
         "systems": {},
+        "subcontracted_equipment": {},
         "fees": {},
         "subtotal": decimal.Decimal(0.00),
+        "subcontracted_amount": decimal.Decimal(0.00),
         "fees_amt": decimal.Decimal(0.00),
         "payment_amt": decimal.Decimal(0.00),
         "total_amt": decimal.Decimal(0.00),
@@ -39,6 +41,7 @@ def calculateGigCost(estimate):
         )
 
         total_loadin_time = decimal.Decimal(0.00)
+        total_show_time = decimal.Decimal((ret["gig"].end - ret["gig"].setup_by) / timezone.timedelta(minutes=15) / 4)
         for loadin in dept_loadins.order_by("shop_time"):
             total_loadin_time += decimal.Decimal(
                 (
@@ -76,7 +79,7 @@ def calculateGigCost(estimate):
         addons = systeminstance.addoninstance_set.all()
         for addon_set_item in addons:
             addon = addon_set_item.addon
-            addon.addl_description = addon_set_item.description
+            addon.description = f"{(f'{addon.description}') if addon.description else ''} {f' - {addon_set_item.description}' if addon_set_item.description else ''}"
 
             current_price = addon.get_price_at_date(ret["gig"].start)
 
@@ -91,6 +94,11 @@ def calculateGigCost(estimate):
                     current_price.price_per_hour_for_load_in_out_ONLY
                     * addon_set_item.qty
                     * total_loadin_time
+                )
+                + (
+                    current_price.price_per_hour_for_show_ONLY
+                    * addon_set_item.qty
+                    * total_show_time
                 ),
                 2,
             )
@@ -101,13 +109,52 @@ def calculateGigCost(estimate):
                 addon_set_item.qty * total_time_rented
                 if current_price.price_per_hour != 0.00
                 else addon_set_item.qty * total_loadin_time
-                if current_price.price_per_hour_for_load_in_out_ONLY
+                if current_price.price_per_hour_for_load_in_out_ONLY != 0.00
+                else addon_set_item.qty * total_show_time
+                if current_price.price_per_hour_for_show_ONLY != 0.00
                 else addon_set_item.qty,
                 addon_subtotal,
                 True,
             ]
         ret["subtotal"] += system_subtotal
         ret["total_amt"] += system_subtotal
+
+    for rental in ret["estimate"].gig.subcontractedequipment_set.all():
+        ret["subcontracted_equipment"][rental.vendor] = {
+            "fees": [],
+            "equipment": {},
+        }
+        vendor_subtotal = decimal.Decimal(0.00)
+        for instance in rental.subcontractedequipmentinstance_set.all():
+            subtotal = decimal.Decimal(0.00)
+            total_time_rented = decimal.Decimal(
+                (rental.returned - rental.arrival) / timezone.timedelta(minutes=15) / 4
+            )
+            current_price = instance.equipment.get_price_at_date(rental.arrival)
+            subtotal += round(
+                current_price.base_price * instance.qty
+                + (
+                    current_price.price_per_hour
+                    * instance.qty
+                    * total_time_rented
+                ), 2
+            )
+            vendor_subtotal += subtotal
+            ret["subcontracted_equipment"][rental.vendor]["equipment"][instance.pk] = [
+                instance,
+                instance.qty * total_time_rented
+                if current_price.price_per_hour != 0.00
+                else instance.qty,
+                current_price.price_per_hour if current_price.price_per_hour != 0.00 else current_price.base_price,
+                subtotal
+            ]
+        for fee in rental.vendorfee_set.all():
+            amount = round(fee.percentage*subtotal+fee.amount, 2);
+            ret["subcontracted_equipment"][rental.vendor]["fees"].append([fee, amount])
+            vendor_subtotal += amount
+
+        ret["subcontracted_amount"] += vendor_subtotal
+        ret["total_amt"] += vendor_subtotal
 
     for fee in ret["estimate"].onetimefee_set.order_by("order").all():
         fee_amt = (
@@ -124,4 +171,45 @@ def calculateGigCost(estimate):
 
     ret["total_amt"] = ret["total_amt"] + ret["estimate"].adjustments
     ret["outstanding_balance"] = ret["total_amt"] - ret["payment_amt"]
+    return ret
+
+
+def calcuateSubcontractedCost(subcontracted):
+    ret = {
+        "subcontracted": subcontracted,
+        "subtotal": decimal.Decimal(0.00),
+        "total_amt": decimal.Decimal(0.00),
+        "fee_amt": decimal.Decimal(0.00),
+        "fees": [],
+        "equipment": {},
+    }
+    for instance in subcontracted.subcontractedequipmentinstance_set.all():
+        subtotal = decimal.Decimal(0.00)
+        total_time_rented = decimal.Decimal(
+            (subcontracted.returned - subcontracted.arrival) / timezone.timedelta(minutes=15) / 4
+        )
+        current_price = instance.equipment.get_price_at_date(subcontracted.arrival)
+        subtotal += round(
+            current_price.base_price * instance.qty
+            + (
+                current_price.price_per_hour
+                * instance.qty
+                * total_time_rented
+            ), 2
+        )
+        ret["total_amt"] += subtotal
+        ret["equipment"][instance] = [
+            instance,
+            instance.qty * total_time_rented
+            if current_price.price_per_hour != 0.00
+            else instance.qty,
+            current_price.price_per_hour if current_price.price_per_hour != 0.00 else current_price.base_price,
+            subtotal
+        ]
+        ret["subtotal"] += subtotal
+    for fee in subcontracted.vendorfee_set.all():
+        amount = round(fee.percentage*subtotal+fee.amount, 2);
+        ret["fees"].append([fee, amount])
+        ret["fee_amt"] += amount
+    ret["total_amt"] += ret["fee_amt"]
     return ret