+++ finance/models.py
@@ -129,11 +129,22 @@ class SystemAddonPricing(BasePricing):
     """
     addon = models.ForeignKey("equipment.SystemAddon", on_delete=models.CASCADE)
     price_per_hour_for_load_in_out_ONLY = models.DecimalField(
-        max_digits=8, decimal_places=2, default=0.00
+        max_digits=8, decimal_places=2, default=0.00, help_text="Price per hour from first dept load in to last dept load out"
+    )
+    price_per_hour_for_show_ONLY = models.DecimalField(
+        max_digits=8, decimal_places=2, default=0.00, help_text="Price per hour from setup by time to end of show"
     )
 
     def clean(self, *args, **kwargs):
         kwargs["item_set"] = self.__class__.objects.filter(addon=self.addon)
+        super().clean(*args, **kwargs)
+
+
+class VendorEquipmentPricing(BasePricing):
+    equipment = models.ForeignKey("equipment.VendorEquipment", on_delete=models.CASCADE)
+
+    def clean(self, *args, **kwargs):
+        kwargs["item_set"] = self.__class__.objects.filter(equipment=self.equipment)
         super().clean(*args, **kwargs)
 
 
@@ -166,6 +177,13 @@ class FeePricing(Pricing):
     def clean(self, *args, **kwargs):
         kwargs["item_set"] = self.__class__.objects.filter(fee=self.fee)
         super().clean(*args, **kwargs)
+
+
+class BaseFee(models.Model):
+    name = models.CharField(max_length=64, blank=True, null=True)
+    amount = models.DecimalField(max_digits=7, decimal_places=2, default=0.00)
+    percentage = models.DecimalField(max_digits=5, decimal_places=2, default=0.00)
+    description = models.CharField(max_length=512, blank=True, null=True)
 
 
 class OneTimeFee(models.Model):
@@ -201,9 +219,18 @@ class OneTimeFee(models.Model):
             self.percentage = pricing.percentage
             self.description = self.prepared_fee.description
         super().save(*args, **kwargs)
+        self.estimate.save()
+
+    def delete(self, *args, **kwargs):
+        super().delete(*args, **kwargs)
+        self.estimate.save()
 
     def __str__(self):
         return f"{self.name} - {f'${self.amount}' if self.amount else f'{self.percentage}%' if self.percentage else ''}"  # noqa
+
+
+class VendorFee(BaseFee):
+    equipment_order = models.ForeignKey("gig.SubcontractedEquipment", on_delete=models.CASCADE)
 
 
 class Estimate(models.Model):
@@ -335,7 +362,7 @@ class Estimate(models.Model):
             )
 
     def __str__(self):
-        return f"{self.get_status_display()} - {self.gig} - ${self.total_amt}"
+        return f"{self.get_status_display()}{f' (OB: ${self.outstanding_balance})' if (self.outstanding_balance != 0 and self.status == 'A') else ''} - {self.gig} - ${self.total_amt}"
 
 
 class CannedNote(models.Model):
@@ -377,8 +404,6 @@ class Shift(models.Model):
 
     def clean(self, *args, **kwargs):
         # add custom validation here
-        if self.content_object.position.hourly_rate is None:
-            raise ValidationError(f"This position does not have an associated hourly rate. Please add one or choose another position.")
         super().clean(*args, **kwargs)
 
     def save(self, *args, **kwargs):
@@ -554,6 +579,14 @@ class Payment(models.Model):
     amount = models.DecimalField(max_digits=7, decimal_places=2)
     payment_type = models.CharField(choices=PAYMENT_TYPES, max_length=1)
     estimate = models.ForeignKey("Estimate", on_delete=models.CASCADE)
+
+    def save(self, *args, **kwargs):
+        super().save(*args, **kwargs)
+        self.estimate.save()
+        
+    def delete(self, *args, **kwargs):
+        super().delete(*args, **kwargs)
+        self.estimate.save()
 
     def __str__(self):
         return f"{self.payment_date} - ${self.amount}"