+++ gig/views.py
@@ -12,11 +12,12 @@ from dev_utils.views import MultipleFormView
 from django.views.generic.base import TemplateView
 from django.views.generic.edit import FormView
 from SLUGS.views import SLUGSMixin, isAdminMixin
-from gig.models import Gig, Job, JobInterest
+from gig.models import Gig, Job, JobInterest, BingoBoard
 from employee.models import Employee
 from datetime import datetime
 from django.db.models import F, DateTimeField, ExpressionWrapper
 from django.utils import timezone
+from django.urls import reverse
 
 
 from finance.forms import ShiftFormSet
@@ -50,6 +51,8 @@ class gigIndex(SLUGSMixin, MultipleFormView):
         return context
 
     def dispatch(self, request, *args, **kwargs):
+        if not request.user.is_authenticated:
+            return redirect("%s?next=%s" % (reverse("login"), request.path))
         self.form_classes = {
             "show_notes": {
                 "form": engineerNotesForm,
@@ -58,7 +61,7 @@ class gigIndex(SLUGSMixin, MultipleFormView):
         jobs = {}
         self.added_context["helper"] = shiftFormHelper()
         self.added_context["gig"] = Gig.objects.get(pk=kwargs["gig_id"])
-        if not self.added_context["gig"].published or not request.user.is_authenticated:
+        if not self.added_context["gig"].published:
             raise PermissionDenied()
         try:
             self.added_context["my_jobs"] = Job.objects.filter(employee=request.user, gig=self.added_context["gig"])
@@ -83,6 +86,11 @@ class gigIndex(SLUGSMixin, MultipleFormView):
             }
         self.form_classes["show_notes"]["instance"] = self.added_context["gig"]
         self.added_context["job_forms"] = jobs
+        if BingoBoard.objects.filter(gig=self.added_context["gig"]).count() < 1:
+            self.added_context["bingo_board"] = BingoBoard(gig=self.added_context["gig"])
+            self.added_context["bingo_board"].save()
+        else:
+            self.added_context["bingo_board"] = BingoBoard.objects.get(gig=self.added_context["gig"])
         return super().dispatch(request, *args, **kwargs)
 
     def process_forms(self, form_instances):